Rename test classes to have Test suffix
